+++ src/commands/build-react.js
@@ -4,12 +4,15 @@ import {UserError} from '../errors'
 import webpackBuild from '../webpackBuild'
 import cleanApp from './clean-app'
 
-// Use a config function, as this won't be called until after NODE_ENV has been
-// set by webpackBuild() and we don't want these optimisations in development
-// builds.
-let buildConfig = (args) => {
+// Using a config function as webpackBuild() sets NODE_ENV to production if it
+// hasn't been set by the user and we don't want production optimisations in
+// development builds.
+function buildConfig(args) {
   let entry = args._[1]
   let dist = args._[2] || 'dist'
+
+  let production = process.env.NODE_ENV === 'production'
+  let filenamePattern = production ? '[name].[chunkhash:8].js' : '[name].js'
 
   let config = {
     babel: {
@@ -22,12 +25,14 @@ let buildConfig = (args) => {
       app: path.resolve(entry),
     },
     output: {
-      filename: '[name].js',
+      filename: filenamePattern,
+      chunkFilename: filenamePattern,
       path: path.resolve(dist),
       publicPath: '/',
     },
     plugins: {
       html: {
+        chunksSortMode: 'none',
         mountId: args['mount-id'] || 'app',
         title: args.title || 'React App',
       },
@@ -45,7 +50,7 @@ let buildConfig = (args) => {
     }
   }
 
-  if (process.env.NODE_ENV === 'production') {
+  if (production) {
     config.babel.presets.push('react-prod')
   }