+++ src/commands/build-web-app.js
@@ -5,27 +5,46 @@ import {copyPublicDir} from '../utils'
 import webpackBuild from '../webpackBuild'
 import cleanApp from './clean-app'
 
-export default function buildWebApp(args, cb) {
+// Using a config function as webpackBuild() sets NODE_ENV to production if it
+// hasn't been set by the user and we don't want production optimisations in
+// development builds.
+function buildConfig(args) {
   let entry = args._[1] || 'src/index.js'
   let dist = args._[2] || 'dist'
 
-  cleanApp({_: ['clean-app', dist]})
+  let production = process.env.NODE_ENV === 'production'
+  let filenamePattern = production ? '[name].[chunkhash:8].js' : '[name].js'
 
-  console.log('nwb: build-web-app')
-  copyPublicDir('public', dist)
-  webpackBuild(args, {
+  return {
     devtool: 'source-map',
     entry: {
       app: path.resolve(entry),
     },
     output: {
-      filename: '[name].js',
+      filename: filenamePattern,
+      chunkFilename: filenamePattern,
       path: path.resolve(dist),
       publicPath: '/',
     },
     plugins: {
-      html: getDefaultHTMLConfig(),
+      html: {
+        ...getDefaultHTMLConfig(),
+        chunksSortMode: 'none',
+      },
       vendorChunkName: 'vendor',
     },
-  }, cb)
+  }
+}
+
+/**
+ * Build a plain JS app.
+ */
+export default function buildWebApp(args, cb) {
+  let dist = args._[2] || 'dist'
+
+  cleanApp({_: ['clean-app', dist]})
+
+  console.log(`nwb: build-web-app`)
+  copyPublicDir('public', dist)
+  webpackBuild(args, buildConfig, cb)
 }